[FIRRTL][HW] Shared InstancePathCache between HW and FIRRTL
