+++ lib/Dialect/FIRRTL/Transforms/EmitOMIR.cpp
@@ -32,6 +32,7 @@
 
 using namespace circt;
 using namespace firrtl;
+using hw::InstancePathCache;
 using mlir::LocationAttr;
 using mlir::UnitAttr;
 
@@ -727,8 +728,8 @@ void EmitOMIRPass::makeTrackerAbsolute(Tracker &tracker) {
     namepath.push_back(getInnerRefTo(op));
   };
   // Add the path up to where the NLA starts.
-  for (InstanceOp inst : paths[0])
-    addToPath(inst, inst.getNameAttr());
+  for (auto inst : paths[0])
+    addToPath(inst, inst.instanceNameAttr());
   // Add the path from the NLA to the op.
   if (tracker.nla) {
     auto path = tracker.nla.getNamepath().getValue();
